Relationship: between user and only his notes

+++ backend/controller/note.controller.js
@@ -1,12 +1,14 @@
 const { NoteModel } = require("../models/note.model");
 
 exports.allNotes = async (req, res) => {
-  const notes = await NoteModel.find();
+  const notes = await NoteModel.find({userID:req.userID});
   res.status(200).send({ msg: notes });
 };
 
 exports.addNote = async (req, res) => {
+  const userID = req.userID;
   const payLoad = req.body;
+  payLoad.userID = userID;
   try {
     const newNote = new NoteModel(payLoad);
     await newNote.save();
@@ -25,6 +27,10 @@ exports.updateNote = async (req, res) => {
   const note = await NoteModel.findOne({ _id });
   if (!note) return res.status(404).send({ msg: "note not found" });
 
+  // check authorized
+  if (note.userID !== req.userID) {
+    return res.status(403).send({ msg: "You are not authorized" });
+  }
   // Check if payload is empty
   if (Object.keys(payLoad).length === 0)
     return res.status(400).send({ msg: "Write something to update" });
@@ -42,7 +48,10 @@ exports.deleteNote = async (req, res) => {
   //check for notes
   const note = await NoteModel.findOne({ _id });
   if (!note) return res.status(404).send({ msg: "note not found" });
-
+  // check authorized
+  if (note.userID !== req.userID) {
+    return res.status(403).send({ msg: "You are not authorized" });
+  }
   try {
     await NoteModel.findOneAndDelete({ _id });
     res.status(200).send({ msg: "Note deleted" });
